Drivers: DMIC: Select FIR decimation factor from preferred order
This patch changes the preference from minimum FIR decimation factor
into introduction order of decimation factors in vector fir_list[]
that is defined in pdm_decim_table.h. The preferred decimation factors
are in this order: 2, 4, 6, 8, 3, 5, 12.

The prime numbers 3 and 5 are less preferred since there is more product
factors for decimation ratios for the FIR for the other FIFO with
these vs. the primes. If the FIR for 16 kHz output would decimate by
factor 5 there is no integer factor to decimate from the 80 kHz rate to
48 kHz. But decimation by 6 enables to decimate by 2 from 96 kHz.

This will needs plenty of testing, also with other than 16 and 48 kHz
where this seems to work OK.

Signed-off-by: Seppo Ingalsuo <seppo.ingalsuo@linux.intel.com>